% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniMINstar.R
\name{UniMINstar}
\alias{UniMINstar}
\title{UniMINstar test for testing equality means against umbrella-ordered alternatives (unknown peak) in one-way ANOVA}
\usage{
UniMINstar(sample_data, significance_level)
}
\arguments{
\item{sample_data}{list}

\item{significance_level}{numeric}
}
\value{
Peak numeric

Critical value numeric

Test statistic value numeric

Result Character
}
\description{
UniMINstar test for testing equality means against umbrella-ordered alternatives (unknown peak) in one-way ANOVA
}
\details{
Testing of H_0:mu_1 = mu_2 = ... = mu_k vs H_1:mu_1 <=.....<= mu_(h-1)<= mu_h >= mu_(h+1)>=....>= mu_k (at least one strict inequality, h is unknown), where mu_i represents the population means of the i-th treatment. The input consists of two variables: sample_data and significance_level. The output consists of the peak of the umbrella, the critical value, the UniMIN* test statistic value, and the result, which indicates whether to reject or not reject the null hypothesis.
}
\author{
Subha Halder
}
